split JsonPath implementation
