+++ be/src/exprs/vectorized/json_functions.h
@@ -17,8 +17,6 @@ DIAGNOSTIC_POP
 namespace starrocks {
 namespace vectorized {
 
-namespace vpack = arangodb::velocypack;
-
 enum JsonFunctionType {
     JSON_FUN_INT = 0,
     JSON_FUN_DOUBLE,
@@ -45,64 +43,7 @@ struct JsonTypeTraits<TYPE_VARCHAR> {
     static JsonFunctionType JsonType;
 };
 
-enum ArraySelectorType {
-    INVALID,
-    NONE,
-    SINGLE,
-    WILDCARD,
-    SLICE,
-};
-
-// Array selector syntax:
-// 1. arr[x] select the x th element
-// 2. arr[*] select all elements
-// 3. arr[1:3] select slice of elements
-struct ArraySelector {
-    ArraySelectorType type = INVALID;
-
-    ArraySelector() {}
-    virtual ~ArraySelector() = default;
-
-    static Status parse(const std::string& str, std::unique_ptr<ArraySelector>* output);
-
-    static bool match(const std::string& input) { return false; }
-
-    virtual void iterate(vpack::Slice array_slice, std::function<void(vpack::Slice)> callback) = 0;
-};
-
-struct ArraySelectorNone final : public ArraySelector {
-    ArraySelectorNone() { type = NONE; }
-
-    virtual void iterate(vpack::Slice array_slice, std::function<void(vpack::Slice)> callback) override { return; }
-};
-
-struct ArraySelectorSingle final : public ArraySelector {
-    int index;
-
-    ArraySelectorSingle(int index) : index(index) { type = SINGLE; }
-
-    static bool match(const std::string& input);
-
-    void iterate(vpack::Slice array_slice, std::function<void(vpack::Slice)> callback) override;
-};
-
-struct ArraySelectorWildcard final : public ArraySelector {
-    ArraySelectorWildcard() { type = WILDCARD; }
-
-    static bool match(const std::string& input);
-
-    void iterate(vpack::Slice array_slice, std::function<void(vpack::Slice)> callback) override;
-};
-
-struct ArraySelectorSlice final : public ArraySelector {
-    int left, right;
-
-    ArraySelectorSlice(int left, int right) : left(left), right(right) { type = SLICE; }
-
-    static bool match(const std::string& input);
-
-    void iterate(vpack::Slice array_slice, std::function<void(vpack::Slice)> callback) override;
-};
+extern const re2::RE2 JSON_PATTERN;
 
 struct SimpleJsonPath {
     std::string key; // key of a json object
@@ -132,22 +73,6 @@ struct SimpleJsonPath {
         ss << "key: " << key << ", idx: " << idx << ", valid: " << is_valid;
         return ss.str();
     }
-};
-
-// JsonPath implement that suuport array building
-struct JsonPath {
-    std::string key;
-    std::unique_ptr<ArraySelector> array_selector;
-
-    JsonPath(const std::string& key, std::unique_ptr<ArraySelector>&& selector)
-            : key(key), array_selector(std::move(selector)) {}
-
-    JsonPath(const std::string& key, ArraySelector* selector) : key(key), array_selector(selector) {}
-
-    JsonPath() = default;
-    JsonPath(JsonPath&& rhs) = default;
-
-    ~JsonPath() = default;
 };
 
 class JsonFunctions {
@@ -229,15 +154,7 @@ public:
 
     static void parse_json_paths(const std::string& path_strings, std::vector<SimpleJsonPath>* parsed_paths);
 
-    static Status parse_full_json_paths(const std::string& path_strings, std::vector<JsonPath>* parsed_paths);
-
 private:
-    static vpack::Slice _extract_full_from_object(const JsonValue* json, const std::vector<JsonPath>& jsonpath,
-                                                  vpack::Builder* b);
-
-    static vpack::Slice _extract_full_from_object(vpack::Slice root, const std::vector<JsonPath>& jsonpath,
-                                                  int path_index, vpack::Builder* b);
-
     // TODO(mofei) deprecated
     static Status _get_parsed_paths(const std::vector<std::string>& path_exprs,
                                     std::vector<SimpleJsonPath>* parsed_paths);
